PRSD-793: Simple review markups
